Started working in fixing the backend
