+++ src/app/api/db/entities/internal-reading/entity.ts
