+++ src/app/api/db/entities/prototype/entry.ts
@@ -1,0 +1,12 @@
+export type PrototypeEntry = {
+    id: string,
+    key: string,
+    operational: boolean,
+    activationCode: string,
+    version: string,
+    owner: {
+        name: string,
+        surname: string,
+        id: string,
+    } | null
+};
