+++ src/app/api/db/entities/user/_user/entity.ts
@@ -1,0 +1,7 @@
+export type _UserEntity = {
+    _id?: string,
+    name: string,
+    surname: string,
+    email: string,
+    encrypted_password: string,
+};
