+++ src/app/api/db/entities/user/_user/input.ts
@@ -1,0 +1,6 @@
+export type _UserRegistrationInput = {
+    name: string,
+    surname: string,
+    email: string,
+    password: string,
+};
