+++ src/app/api/db/entities/user/admin/entity.ts
@@ -1,0 +1,12 @@
+import { _UserEntity } from "../_user/entity";
+
+export type AdminLabel = "admin";
+
+export type AdminPermissions = "read" | "all";
+
+export type AdminEntity = _UserEntity & {
+    type: AdminLabel,
+    admin_code: string,
+    permissions: AdminPermissions,
+    invited_admins: string[],
+};
