+++ src/app/api/db/entities/user/admin/entry.ts
@@ -1,0 +1,10 @@
+import { AdminPermissions } from "./entity";
+
+export type AdminEntry = {
+    id: string,
+    name: string,
+    surname: string,
+    email: string,
+    adminCode: string,
+    permissions: AdminPermissions,
+};
