+++ src/app/api/db/entities/user/admin/input.ts
@@ -1,0 +1,8 @@
+import { _UserRegistrationInput } from "../_user/input";
+import { AdminLabel } from "./entity";
+
+export type AdminRegistrationInput = _UserRegistrationInput & {
+    type: AdminLabel,
+    adminEmail: string,
+    adminCode: string,
+};
