+++ src/app/api/db/entities/user/customer/entity.ts
@@ -1,0 +1,8 @@
+import { _UserEntity } from "../_user/entity";
+
+export type CustomerLabel = "customer";
+
+export type CustomerEntity = _UserEntity & {
+    type: CustomerLabel,
+    prototypes: string[]
+};
