+++ src/app/api/db/entities/user/customer/entry.ts
@@ -1,0 +1,6 @@
+export type CustomerEntry = {
+    id: string,
+    name: string,
+    surname: string,
+    email: string
+}
