+++ src/app/api/db/entities/user/customer/preview.ts
@@ -1,0 +1,14 @@
+import { NotificationPreview } from "../../notification/preview";
+import { PrototypePreview } from "../../prototype/preview";
+import { CustomerLabel } from "./entity";
+
+export type CustomerPreview = {
+    id: string;
+    name: string;
+    surname: string;
+    email: string;
+    type: CustomerLabel;
+    notifications: NotificationPreview[];
+    oldestNotification: string | null;
+    prototypes: PrototypePreview[];
+};
