+++ src/app/api/notification/all/route.ts
@@ -6,7 +6,7 @@ import { query, where, getDocsFromServer } from "firebase/firestore";
 import { makeErrorResponse } from "../../lib/make-error-response";
 import { verifyJwt } from "@/app/api/lib/jwt";
 import { parseEntity } from "../../db/parse-entity";
-import { NotificationEntity } from "../../db/entities/notification-entity";
+import { NotificationEntity } from "../../db/entities/notification/entity";
 
 export async function GET(request: NextRequest) {
     try {
